chagne: rename pennylane base class name
