+++ qxmt/feature_maps/pennylane/__init__.py
@@ -1,15 +1,15 @@
+from qxmt.feature_maps.pennylane.base import PennyLaneBaseFeatureMap
 from qxmt.feature_maps.pennylane.ising import XXFeatureMap, YYFeatureMap, ZZFeatureMap
 from qxmt.feature_maps.pennylane.npqc import NPQCFeatureMap
-from qxmt.feature_maps.pennylane.pennylane_base import PennyLaneBaseFeatureMap
 from qxmt.feature_maps.pennylane.rotation import HRotationFeatureMap, RotationFeatureMap
 from qxmt.feature_maps.pennylane.yzcx import YZCXFeatureMap
 
 __all__ = [
+    "PennyLaneBaseFeatureMap",
     "XXFeatureMap",
     "YYFeatureMap",
     "ZZFeatureMap",
     "NPQCFeatureMap",
-    "PennyLaneBaseFeatureMap",
     "HRotationFeatureMap",
     "RotationFeatureMap",
     "YZCXFeatureMap",
